Pass physicsdata through dycore

+++ jcm/date.py
@@ -161,32 +161,32 @@ class Timestamp:
 @tree_math.struct
 class DateData:
     tyear: jnp.float32 # Fractional time of year, should possibly be part of the model itself (i.e. not in physics_data)
-    model_year: jnp.int32
-    model_step: jnp.int32
+    model_year: jnp.float32 # FIXME: should be integer
+    model_step: jnp.float32 # FIXME: should be integer
 
     @classmethod
     def zeros(self, model_time=None, model_year=None, model_step=None):
         return DateData(
           tyear=fraction_of_year_elapsed(model_time) if model_time is not None else 0.0,
           model_year=model_year if model_year is not None else 1950,
-          model_step=model_step if model_step is not None else jnp.int32(0))
+          model_step=model_step if model_step is not None else jnp.float32(0))
 
     @classmethod
     def set_date(self, model_time, model_year=None, model_step=None):
         return DateData(
           tyear=fraction_of_year_elapsed(model_time),
           model_year=model_year if model_year is not None else 1950,
-          model_step=model_step if model_step is not None else jnp.int32(0))
+          model_step=model_step if model_step is not None else jnp.float32(0))
 
     @classmethod
     def ones(self, model_time=None, model_year=None, model_step=None):
         return DateData(
           tyear=fraction_of_year_elapsed(model_time) if model_time is not None else 1.0,
           model_year=model_year if model_year is not None else 1950,
-          model_step=model_step if model_step is not None else jnp.int32(0))
+          model_step=model_step if model_step is not None else jnp.float32(0))
 
     def model_day(self):
-        return jnp.round(self.tyear*days_year).astype(jnp.int32)
+        return jnp.round(self.tyear*days_year).astype(jnp.float32)
 
     def copy(self, tyear=None, model_year=None, model_step=None):
         return DateData(
